Tools entities

+++ domain/src/main/kotlin/com.nstu.technician.domain/model/tool/Component.kt
@@ -1,7 +1,7 @@
 package com.nstu.technician.domain.model.tool
 
 class Component(
-    val oid: Int,
+    val oid: Long,
     val name: String,
     val componentType: ComponentType
 )
